requirements.txt
attrs==21.4.0
jsonmerge==1.8.0
jsonschema==4.5.1
pyrsistent==0.18.1
pyserial==3.5
websocket-client==1.3.2
